pyproject.toml
[tool.poetry]
name = "djangofloors"
version = "0.1.0"
description = ""
authors = ["nail <nail20142511@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
Django = "^3.1.5"
Pillow = "^8.1.0"
django-fieldsets-with-inlines = "^0.6"
pymorphy2 = "^0.9.1"
django-debug-toolbar = "^3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
